+++ src/common/components/Input/components/Description/index.tsx
@@ -1,5 +1,6 @@
+import { TextBoxProps } from '@components/Input/types';
+
 import { descriptionVar } from './style.css';
-import { TextBoxProps } from '../types';
 
 // TextBox 내부 Input 하단의 부가텍스트
 const Description = ({ children, styleType = 'default' }: Pick<TextBoxProps, 'children' | 'styleType'>) => (
