+++ src/common/components/Input/components/InputButton/index.tsx
@@ -1,7 +1,7 @@
 import Button from '@components/Button';
 
 import { textWidth } from './style.css';
-import { InputButtonProps } from '../types';
+import { InputButtonProps } from './types';
 
 // TextBox 내부 InputLine 우측 버튼
 const InputButton = ({ isLoading, text, ...props }: InputButtonProps) => {
