+++ src/common/components/Input/components/Timer/types.ts
@@ -1,0 +1,4 @@
+export interface TimerProps {
+  isActive: boolean;
+  onResetTimer: () => void;
+}
