+++ src/common/components/Layout/components/Footer/icons/IconArrowRight.tsx
