+++ src/common/components/Layout/components/Footer/icons/IconMail.tsx
